MOVE: HttpRPC to slatekit.http module
