+++ src/lib/kotlin/slatekit-http/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'slatekit-http'
