+++ src/lib/kotlin/slatekit-notifications/src/main/kotlin/slatekit/notifications/push/PushGoogle.kt
@@ -14,9 +14,9 @@
 package slatekit.notifications.push
 
 import okhttp3.Request
-import slatekit.common.*
 import slatekit.common.conf.Conf
 import slatekit.common.log.Logs
+import slatekit.http.HttpRPC
 import slatekit.notifications.common.Sender
 import slatekit.results.Outcome
 import slatekit.results.Success
