+++ src/lib/kotlin/slatekit-notifications/src/main/kotlin/slatekit/notifications/sms/SmsServiceTwilio.kt
@@ -14,10 +14,10 @@
 package slatekit.notifications.sms
 
 import okhttp3.Request
-import slatekit.common.*
 import slatekit.common.info.ApiLogin
 import slatekit.common.templates.Templates
 import slatekit.common.types.CountryCode
+import slatekit.http.HttpRPC
 import slatekit.results.*
 import slatekit.results.builders.Outcomes
